wic: Add a specific wks file for this BSP
Issue: LIN10-3263

The sdimage-bootpart.wks in the oe-core layer uses the "mmcblk" as
the mmc block device. This would cause the WIC to add a mount entry
like below in the fstab:
    /dev/mmcblkp1	/boot	vfat	defaults	0	0

Then we would get stuck when trying to mount it. So add a specific wks
file for this BSP to fix this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>