nxp-ls20xx: add the command line condition to use Kexec
Issue: LIN1018-3468

The issue is that the PCIe Ethernet does not work on Kexec.

For Kexec, the second kernel can not use the MSI of PCI since the disabling
LPIs of GIC3 failure, Enabling LPIs of GIC3 is a one way switch, once they're on,
there is no turning back. On WRLinux 10.18 kernel, added the judgement for
disabling LPIs successful but not on WRLinux 10.17. So lucky, it worked fine
on WRLinux 10.17, but the system is horribly unsafe on Kexec.

Here add the command line condition in the README.

Detailed reason for Kexec about disabling LPI could refer to the upstream commit:

f736d65df0a (irqchip/gic-v3: Allow LPIs to be disabled from the command line)

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>